app/requirements.txt
fastapi==0.105.0
uvicorn[standard]==0.23.0
python-telegram-bot==22.2
python-dotenv==1.0.0
loguru==0.7.3
slowapi==0.1.9
jinja2==3.1.2
pydantic
sqlalchemy==2.0.44
psycopg2-binary
PyJWT==2.8.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
aiofiles==23.1.0
